Start refactor to use new source plugin
