+++ .npmrc
@@ -1,4 +1,2 @@
+# We don't need Artifactory at the moment, so explicitly setting default npm registry
 registry=https://artifactory.corp.mongodb.com/artifactory/api/npm/npm/
-//artifactory.corp.mongodb.com/artifactory/api/npm/npm/:_auth=${NPM_BASE_64_AUTH}
-_email=${NPM_EMAIL}
-_always-auth=true
