+++ src/context/version-context.js
@@ -173,6 +173,7 @@ const VersionContextProvider = ({ repoBranches, associatedReposInfo, isAssociate
   const [availableGroups, setAvailableGroups] = useState(
     getDefaultGroups(metadata.project, repoBranches, associatedReposInfo)
   );
+  console.log({repoBranches})
   const [showEol, setShowEol] = useState(repoBranches['branches']?.some((b) => !b.active) || false);
 
   // on init, fetch versions from realm app services
