+++ src/hooks/use-remote-metadata.js
@@ -2,16 +2,17 @@ import { useStaticQuery, graphql } from 'gatsby';
 
 // Return the remote metadata node
 export const useRemoteMetadata = () => {
-  const data = useStaticQuery(
-    graphql`
-      query RemoteMetadata {
-        allRemoteMetadata {
-          nodes {
-            remoteMetadata
-          }
-        }
-      }
-    `
-  );
-  return data.allRemoteMetadata.nodes[0].remoteMetadata;
+  // const data = useStaticQuery(
+    // graphql`
+      // query RemoteMetadata {
+        // allRemoteMetadata {
+          // nodes {
+            // remoteMetadata
+          // }
+        // }
+      // }
+    // `
+  // );
+  // return data.allRemoteMetadata.nodes[0].remoteMetadata;
+  return null
 };
