+++ src/hooks/useAllProducts.js
@@ -2,17 +2,18 @@ import { useStaticQuery, graphql } from 'gatsby';
 
 // Return an array of MongoDB products
 export const useAllProducts = () => {
-  const { allProduct } = useStaticQuery(
-    graphql`
-      query AllProducts {
-        allProduct {
-          nodes {
-            title
-            url
-          }
-        }
-      }
-    `
-  );
-  return allProduct.nodes;
+  // const { allProduct } = useStaticQuery(
+    // graphql`
+      // query AllProducts {
+        // allProduct {
+          // nodes {
+            // title
+            // url
+          // }
+        // }
+      // }
+    // `
+  // );
+  // return allProduct.nodes;
+  return null
 };
