+++ src/utils/use-changelog-data.js
@@ -1,15 +1,16 @@
 import { useStaticQuery, graphql } from 'gatsby';
 
 export default function useChangelogData() {
-  const data = useStaticQuery(graphql`
-    query ChangelogData {
-      allChangelogData {
-        nodes {
-          changelogData
-        }
-      }
-    }
-  `);
+  // const data = useStaticQuery(graphql`
+    // query ChangelogData {
+      // allChangelogData {
+        // nodes {
+          // changelogData
+        // }
+      // }
+    // }
+  // `);
 
-  return data.allChangelogData.nodes[0].changelogData;
+  // return data.allChangelogData.nodes[0].changelogData;
+  return null
 }
